Replace AddContentOverlay with ChangeContentOverlay.
According to the documentation, AddContentOverlay blows away the analysis_servers knowledge of the file and replaces it with the provided contents. However, when we use this in practice the analysis_server seems to completely forget about non-essential (hints, infos, warnings) errors. So now we do the same thing, but by using the ChangeContentOverlay instead.

Closes #74 (and several strange, unreported anamolies I noticed locally)