Resolve an issue related to call order of consumeLinter.
Pre Atom 1.0, consumeLinter was called after the module was activated. In 1.0, however, it looks like consumeLinter can be called _before_ the package is activated, resulting in several expected instance variables to be null.

Therefore, we check whether we've booted or not first: if so, setup the linter immediately. If not, defer execution of the linter until the boot procedure is complete.